+++ assets/js/theme-ui.js
@@ -1,0 +1,10 @@
+jQuery(function ($) {
+
+    $('#navigation a').click(function () {
+
+        $('#navigation .active').removeClass('active');
+        $(this).addClass('active');
+
+    });
+
+});