+++ template-notes.php
@@ -1,4 +1,5 @@
 <?php
+if ( ! defined( 'ABSPATH' ) ) { exit; }
 /*Template Name: Notes*/
 
 get_header();